Revert "Jls dev jan15 (#65)"
This reverts commit f3591c562cc57af271aba8a14536146662feb4f3.
