+++ netcad/cli/netcad/__init__.py
@@ -5,8 +5,11 @@ from .netcad import script
 from . import cli_build_configs
 from . import cli_build_checks
 from . import cli_design_list
+from .. import cli_get
 from . import cli_show_cabling
 from . import cli_report_devices
 from . import cli_diagram
 from . import cli_report_interfaces
 from . import cli_report_ipaddrs
+from . import cli_show_switchports
+from . import cli_show_vlans
