+++ netcad/cli/netcad/netcad.py
@@ -13,7 +13,7 @@ import sys
 
 from netcad.logger import get_logger
 from netcad.debug import format_exc_message
-from netcad.init import init, init_plugings, builtin_plugins
+from netcad.init import init, init_plugings
 from .cli_netcad_main import cli
 
 # -----------------------------------------------------------------------------
@@ -30,7 +30,6 @@ def script():
     """
     try:
         init()
-        builtin_plugins.init_netcad_builtin_plugins()
         init_plugings.init_netcad_plugins()
         cli()
