+++ netcad/device/__init__.py
@@ -6,4 +6,6 @@ from .device_group import PseudoDevice, DeviceGroup, DeviceGroupMember
 from .device_group_mlag import DeviceMLagPairGroup, DeviceMLagPairMember
 from .peer_interface_id import PeerInterfaceId
 from .interface_profile import InterfaceProfile
+from .l2_interfaces import InterfaceL2, InterfaceL2Access, InterfaceL2Trunk
+from .l3_interfaces import InterfaceVlan
 from .lag_interfaces import InterfaceLag, InterfaceLagMember
