+++ netcad/device/l3_interfaces.py
@@ -5,13 +5,15 @@
 # System Imports
 # -----------------------------------------------------------------------------
 
-from typing import Optional
+from typing import Optional, Set
 from ipaddress import IPv4Address
 
 # -----------------------------------------------------------------------------
 # Private Imports
 # -----------------------------------------------------------------------------
 
+from netcad.vlan import vlan_profile as vp
+from netcad.vlan.vlan_profile import VlanProfile
 from .interface_profile import InterfaceVirtual, InterfaceProfile
 
 # -----------------------------------------------------------------------------
@@ -20,6 +22,7 @@ from .interface_profile import InterfaceVirtual, InterfaceProfile
 
 __all__ = [
     "InterfaceL3",
+    "InterfaceVlan",
     "InterfaceIsLoopback",
     "InterfaceIsManagement",
     "InterfaceIsInVRF",
@@ -57,3 +60,15 @@ class InterfaceIsLoopback(InterfaceVirtual, InterfaceL3):
 class InterfaceIsInVRF(InterfaceL3):
     is_in_vrf = True
     vrf: Optional[str] = "management"
+
+
+class InterfaceVlan(InterfaceVirtual, InterfaceL3):
+    """
+    InterfaceVlan is used to declare IP addresses assigned to VLANs.  Also
+    referred to an SVI.
+    """
+
+    vlan: VlanProfile
+
+    def vlans_used(self) -> Set[vp.VlanProfile]:
+        return {self.vlan}
