+++ netcad/vlan/descriptor_vlan_all.py
@@ -22,7 +22,7 @@ from operator import attrgetter
 
 from netcad.device import DeviceInterface
 from netcad.device.interface_profile import InterfaceProfile
-from netcad.vlans import VlanProfile
+from netcad.vlan import VlanProfile
 
 # -----------------------------------------------------------------------------
 # Exports
