--- netcad/vlans/cli/__init__.py
@@ -1,5 +1,0 @@
-#  Copyright (c) 2021 Jeremy Schulman
-#  GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
-
-from . import cli_show_switchports
-from . import cli_show_vlans
