--- netcad/vlans/profiles/__init__.py
@@ -1,5 +1,0 @@
-#  Copyright (c) 2021 Jeremy Schulman
-#  GNU General Public License v3.0+ (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
-
-from .l2_interfaces import InterfaceL2, InterfaceL2Access, InterfaceL2Trunk
-from .vlan_interface import InterfaceVlan
